requirements.txt
# =============================================================================
# Requirements for Data Preparation Pipeline
# =============================================================================

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Configuration
omegaconf>=2.3.0
pyyaml>=6.0

# Visualization (EDA)
matplotlib>=3.7.0
seaborn>=0.13.0

# Progress bar
tqdm>=4.65.0

# HTTP requests (image downloading)
requests>=2.31.0

# Text Extraction - Google Vision API
google-cloud-vision

# Text Extraction - GLM-OCR
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
